requirements.txt

python-dotenv==1.0.1


PyPDF2==3.0.1


langchain==0.1.17
langchain-core==0.1.46


langchain-huggingface==0.0.1
sentence-transformers==2.6.1
transformers==4.41.1


pinecone-client==3.2.2
langchain-pinecone==0.0.3


google-generativeai==0.5.4
langchain-google-genai==0.0.11


gradio==4.27.0


tqdm==4.66.4
